requirements.txt
Package            Version
------------------ -----------
blis               0.4.1
catalogue          1.0.0
certifi            2020.4.5.2
chardet            3.0.4
click              7.1.2
cymem              2.0.3
Django             2.0
en-core-web-sm     2.2.5
filetype           1.0.7
idna               2.9
image              1.5.32
importlib-metadata 1.6.1
joblib             0.15.1
murmurhash         1.0.2
nltk               3.5
numpy              1.18.5
opencv-python      4.2.0.34
pdfminer3          2018.12.3.0
Pillow             7.1.2
pip                20.1.1
plac               1.1.3
preshed            3.0.2
pycryptodome       3.9.7
pytesseract        0.3.4
pytz               2020.1
regex              2020.6.8
requests           2.23.0
setuptools         47.1.1
six                1.15.0
sortedcontainers   2.2.2
spacy              2.2.4
srsly              1.0.2
tesseract          0.1.3
thinc              7.4.0
tqdm               4.46.1
urllib3            1.25.9
wasabi             0.6.0
wheel              0.34.2
zipp               3.1.0
